Actually fix for real this time
The script was failing to click the Compose button everywhere except the one ticket I tested the script on. It turns out that the ID that I was using before to find the button is different for every ticket. Why is it different? HOW is it different? I don't know. This new method for finding the button is slower and uglier but at least it will work now.