Speedup: paste messages instead of typing them
Typing in Cherwell is limited to what feels like 20 WPM because Cherwell. This makes sending keyboard inputs to type out the email you're resending way slower than it needs to be. I reworked the script, and now it uses your clipboard to paste the message into the compose field instantly.