Oh wait, Cherwell can almost resend emails by itself!
I have made a discover that significantly improves this script's speed and stability: the Reply-All function in Cherwell re-writes the email of the selected journal note. There are some things you have to fix to make it look more like an email sent from the regular "click the customer's email address" way, and it is HARD to click this new button automatically, but I've figured it out and now this script runs way faster and works better. No more use of the clipboard and no more text parsing. I'm happy.